have client net methods in a different class
