+++ growtopia/clients/client.py
@@ -2,23 +2,14 @@ __all__ = ("Client",)
 
 import asyncio
 from abc import ABC, abstractmethod
-from typing import Union
-
-import enet
 
 from ..context import ClientContext
 from ..dispatcher import Dispatcher
 from ..enums import EventID
-from ..protocol import (
-    GameMessagePacket,
-    GameUpdatePacket,
-    HelloPacket,
-    StrPacket,
-    TextPacket,
-)
+from .client_net import ClientNet
 
 
-class Client(Dispatcher, ABC):
+class Client(ClientNet, Dispatcher, ABC):
     """
     Represents a client.
     This class can also be used as a base class for other types of clients (e.g game client, proxy client (redirects packets to server)).
@@ -41,76 +32,10 @@ class Client(Dispatcher, ABC):
     """
 
     def __init__(self, address: tuple[str, int], **kwargs) -> None:
-        self._host = enet.Host(
-            None,
-            kwargs.get("peer_count", 1),
-            kwargs.get("channel_limit", 2),
-            kwargs.get("incoming_bandwidth", 0),
-            kwargs.get("outgoing_bandwidth", 0),
-        )
+        ClientNet.__init__(self, address, **kwargs)
         Dispatcher.__init__(self)
 
-        self._host.compress_with_range_coder()
-        self._host.checksum = enet.ENET_CRC32
-
         self.running: bool = False
-
-        self._address: tuple[str, int] = address
-        self._peer: enet.Peer = None
-
-    def connect(self) -> enet.Peer:
-        """
-        Connects to the server.
-
-        Returns
-        -------
-        enet.Peer
-            The peer that was used to connect to the server.
-        """
-        self._peer = self._host.connect(enet.Address(*self._address), 2, 0)
-        return self._peer
-
-    def disconnect(self, send_quit: bool = True) -> None:
-        """
-        Disconnects from the server.
-
-        Parameters
-        ----------
-        send_quit: bool
-            Whether to send the quit packet to the server.
-
-        Returns
-        -------
-        None
-        """
-        if self._peer is None:
-            return
-
-        if send_quit:
-            self.send(packet=TextPacket("action|quit_to_exit\n"))
-            self.send(packet=GameMessagePacket("action|quit\n"))
-
-            self.running = False
-
-        self._peer.disconnect_now(0)
-        self._peer = None
-
-    def send(self, packet: Union[StrPacket, GameUpdatePacket, HelloPacket]) -> bool:
-        """
-        Sends a packet to the host.
-
-        Parameters
-        ----------
-        packet: Union[StrPacket, GameUpdatePacket, HelloPacket]
-            The packet to send to the host.
-
-        Returns
-        -------
-        bool:
-            True if the packet was successfully sent, False otherwise.
-        """
-        if self._peer is not None:
-            return True if self._peer.send(0, packet.enet_packet) == 0 else False
 
     def start(self) -> None:
         """
